requirements.txt
connexion[swagger-ui]
connexion>=2.7.0
Flask_Cors==3.0.8
mysql-connector-python~=8.0.20
requests~=2.23.0
beautifulsoup4==4.9.0
PyJWT==1.7.1
bcrypt==3.1.7
six>=1.14.0
Werkzeug~=1.0.1